% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STBP-eval.R
\name{int_eval}
\alias{int_eval}
\title{Simulation of an STBP}
\usage{
int_eval(pop_mean, prior, n, obj, overdispersion.sim = NA, seed = NULL)
}
\arguments{
\item{pop_mean}{True population density.}

\item{prior}{Initial prior.}

\item{n}{Sample size within bouts.}

\item{obj}{A STBP object}

\item{overdispersion.sim}{Overdispersion parameter or function for simulations.}

\item{seed}{Optional seed for random count generation.}
}
\value{
A list with average number of bouts required to reach a decision and recommendation for H
}
\description{
Simulation of an STBP
}
