% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPRT-eval.R
\name{int_eval2}
\alias{int_eval2}
\title{Simulation of an SPRT}
\usage{
int_eval2(pop_mean, obj, overdispersion.sim = NA, seed = NULL)
}
\arguments{
\item{pop_mean}{True population density.}

\item{obj}{A SPRT object}

\item{overdispersion.sim}{Overdispersion parameter or function for simulations.}

\item{seed}{Optional seed for random count generation.}
}
\value{
A list with average number of bouts required to reach a decision and
recommendation for H1
}
\description{
This function is called on Evaluation of Sequential probability ratio test,
\code{\link{SPRT.eval}}.
}
