% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STBP.R
\name{stbp_posterior_simple}
\alias{stbp_posterior_simple}
\title{Posterior calculation for simple hypotheses}
\usage{
stbp_posterior_simple(
  data,
  hypothesis,
  likelihood_func,
  prior,
  upper_bnd = Inf
)
}
\arguments{
\item{data}{Vector with count data for a single sampling bout.}

\item{hypothesis}{Single value with the hypothesized value of \eqn{\mu}.}

\item{likelihood_func}{Kernel probability density function for the data. See details.}

\item{prior}{Single number with initial prior. Must be in the interval \eqn{[0,1]}}

\item{upper_bnd}{Single number indicating the greatest possible value for \eqn{\mu}. Most cases is \code{Inf}.}
}
\value{
A single probability value
}
\description{
This function calculates a posterior probability for hypotheses about population
densities, of the form \eqn{H:\mu = \psi}, given the data at a single
iteration. This function is to be used in a sequential framework, and called
on the sequential test \code{\link{stbp_simple}}.
}
\details{
The \code{likelihood_func} argument should be specified as a density \code{function} with an \dQuote{x} in the
 mean argument and \dQuote{data} in the vector of quantiles.
 For example, \code{function(data, x) dpois(data, lambda = x)}.
}
\examples{
# Counts collected in a single sampling bout
counts <- c(0, 0, 0)

# Calculate posterior probability from a naive 0.5 prior for H:mu=0
# (a species being absent in an area) with a poisson kernel.

stbp_posterior_simple(data = counts,
                       hypothesis = 0,
                       likelihood_func = function(data, x)
                                                dpois(data, lambda = x),
                       prior = 0.5,
                       upper_bnd = Inf) # returns 0.75
}
\references{
Rincon, D.F., McCabe, I. & Crowder, D.W. (2025) Sequential testing of complementary hypotheses about population density.
\emph{Methods in Ecology and Evolution}. In print.
}
