% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STBP.R
\name{stbp_composite}
\alias{stbp_composite}
\title{Sequential test of Bayesian posterior probabilities for composite hypotheses}
\usage{
stbp_composite(
  data,
  hypothesis,
  likelihood_func,
  prior = 0.5,
  lower_bnd = 0,
  upper_bnd = Inf,
  lower_criterion = 0.01,
  upper_criterion = 0.99
)
}
\arguments{
\item{data}{Either a vector (for purely sequential designs) o a matrix
(group sequential designs) with sequential count data, with sampling bouts
collected over time in columns and sampling within bouts in rows.}

\item{likelihood_func}{Kernel probability density function for the data. See details.}

\item{prior}{Single number with initial prior. Must be in the interval \eqn{[0,1]}.}

\item{lower_bnd}{Single number indicating the lowest possible value for \eqn{\mu}. Most cases is \eqn{0}.}

\item{upper_bnd}{Single number indicating the greatest possible value for \eqn{\mu}. Most cases is \code{Inf}.}

\item{lower_criterion}{Criterion to decide against the tested hypothesis.
This is the lowest credibility to the hypothesis to stop sampling and decide against.}

\item{upper_criterion}{Criterion to decide in favor of the tested hypothesis.
This is the greatest credibility to the hypothesis to stop sampling and decide in favor.}

\item{hypotheses}{Either a single value or a vector with the hypothesized values for \eqn{\mu}.
If a vector, should contain at least as many values as \code{ncol(data)}}
}
\value{
A list with a vector of posterior \code{$probabilities} of length
\code{ncol(data)}, \code{num_iterations} with the number of iterations
required to reach a decision, and a \code{$recommendation}, either
\dQuote{reject H}, \dQuote{accept H}, or \dQuote{keep sampling} for a decision in favor,
against or inconclusive about the hypothesis. The test stops whenever the posterior
probability reached either of the upper or lower criteria.
}
\description{
Runs a Sequential test of Bayesian Posterior Probabilities for hypotheses
about population densities of the form \eqn{H:\mu > \psi}.
Data is treated in a sequential framework.
}
\examples{
# Testing the hypothesis of a population size being greater than 5 individuals
# per sampling unit (H: mu > 5). The sequential sampling is made of 5 sampling
# bouts made of one sample each.

set.seed(101)
counts3 <- rpois(5, lambda = 3)

stbp_composite(data = counts3,
                hypothesis = 5,
                likelihood_func = function(data, x)
                    {dpois(data, lambda = x)},
                prior = 0.5,
                lower_bnd = 0,
                upper_bnd = Inf,
                lower_criterion = 0.001,
                upper_criterion = 0.999)$recommendation
# returns "reject H" after 2 iterations.

# Testing the hypothesis of a sampled population being greater than trajectory H
H <- c(2, 5, 10, 20, 40, 40, 20, 10, 5, 2)

# Generating sequential samples (n = 3) from a population that is 1 below H
# (H - 1)

countP <- matrix(NA, 3, 10)
set.seed(101)
for(i in 1:10){
  countP[, i] <- rpois(3, lambda = (H[i] - 1))
}

# Running STBP on the sample

stbp_composite(data = countP,
                hypothesis = H,
                likelihood_func = function(data, x)
                    {dpois(data, lambda = x)},
                prior = 0.5,
                lower_bnd = 0,
                upper_bnd = Inf,
                lower_criterion = 0.001,
                upper_criterion = 0.999)$recommendation
# returns "reject H" after 8 iterations.
}
