% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STBP.R
\name{stbp_simple}
\alias{stbp_simple}
\title{Sequential test of Bayesian posterior probabilities for simple hypotheses}
\usage{
stbp_simple(
  data,
  hypothesis,
  likelihood_func,
  prior = 0.5,
  upper_bnd = Inf,
  lower_criterion,
  upper_criterion
)
}
\arguments{
\item{data}{Either a vector (for purely sequential designs) o a matrix
(group sequential designs) with sequential count data, with sampling bouts
collected over time in columns and sampling within bouts in rows.}

\item{hypothesis}{Either a single value or a vector with the hypothesized values for \eqn{\mu}.
If a vector, should contain at least as many values as \code{ncol(data)}}

\item{likelihood_func}{Kernel probability density function for the data. See details.}

\item{prior}{Single number with initial prior. Must be in the interval \eqn{[0,1]}.}

\item{upper_bnd}{Single number indicating the greatest possible value for \eqn{\mu}. Most cases is \code{Inf} (default).}

\item{lower_criterion}{Criterion to decide against the tested hypothesis.
This is the lowest credibility to the hypothesis to stop sampling and decide against.}

\item{upper_criterion}{Criterion to decide in favor of the tested hypothesis.
This is the greatest credibility to the hypothesis to stop sampling and decide in favor.}
}
\value{
An object of class \code{"STBP"}.
}
\description{
Runs a Sequential test of Bayesian Posterior Probabilities for hypotheses about population densities of the form \eqn{H:\mu = \psi}.
Data is treated in a sequential framework.
}
\details{
The \code{likelihood_func} argument should be specified as a density \code{function}
with an \dQuote{x} in the mean argument and \dQuote{data} in the vector of quantiles.
For example, \code{function(data, x) dpois(data, lambda = x)}.
}
\examples{

# Testing the absence of a species in a given area from a sequential random
# sampling of 3 bouts made of 10 samples (counts) each (all absences). Upper
# criterion set to 0.9999

counts10 <- matrix(rep(0, 30), 10, 3)

test1G <- stbp_simple(data = counts10,
                        hypothesis = 0,
                        likelihood_func= function(data, x)
                            dpois(data, lambda = x),
                        prior = 0.5,
                        upper_bnd = Inf,
                        lower_criterion = 0,
                        upper_criterion = 0.9999)
test1G

# returns a recommendation of "keep sampling" due to insufficient evidence.

# Testing the same hypothesis with the same upper criterion but from a
# sequential random sampling of 3 bouts made of 30 samples (counts) each
# (all absences).

counts30 <- matrix(rep(0, 90), 30, 3)

test2G <- stbp_simple(data = counts30,
                        hypothesis = 0,
                        likelihood_func= function(data, x)
                            {dpois(data, lambda = x)},
                        prior = 0.5,
                        upper_bnd = Inf,
                        lower_criterion = 0,
                        upper_criterion = 0.9999)
test2G

# returns a recommendation of "accept H" of the species being absent from
# that area.


}
\references{
Rincon, D.F., McCabe, I. & Crowder, D.W. (2025) Sequential testing of complementary hypotheses about population density.
\emph{Methods in Ecology and Evolution}. In print.
}
